Update App.test.js
When I comment out this line, the first test sometimes fails.

Maybe, because by the time the first test finishes, the cart loads in time sometimes and sometimes it doesn't. If the mock server receives a get request for the cart endpoint, it sends back data that also corresponds to the expectation of the first test. Sometimes the cart loads in time, sometimes it doesn't. So, sometimes the expectations are there, sometimes they are not.